# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "API для управления задачами",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список задач",
                "description": "Получение списка задач с возможностью фильтрации",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу задачи",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Фильтр по точной дате выполнения (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "due_date_after",
                        "in": "query",
                        "description": "Фильтр по дате после указанной (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "due_date_before",
                        "in": "query",
                        "description": "Фильтр по дате до указанной (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по названию или описанию",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCollectionResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create new task",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить информацию о задаче",
                "description": "Получение информации о конкретной задаче",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о задаче",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "description": "Обновление задачи",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "description": "Удаление задачи",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача успешно удалена"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "401": {
                        "description": "Требуется авторизация"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreTaskRequest": {
                "required": [
                    "title",
                    "description",
                    "due_date",
                    "status"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Test task"
                    },
                    "description": {
                        "type": "string",
                        "example": "Test description"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-04-25"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "completed"
                        ],
                        "example": "new"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateTaskRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Updated task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated description"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-05-01"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "new",
                            "in_progress",
                            "completed"
                        ],
                        "example": "completed"
                    }
                },
                "type": "object"
            },
            "TaskCollectionResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskResource"
                        }
                    },
                    "meta": {
                        "properties": {
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "required": [
                    "id",
                    "title",
                    "due_date",
                    "status",
                    "user",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Test task"
                    },
                    "description": {
                        "type": "string",
                        "example": "Test description",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-04-25"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-25 21:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-25 21:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}